New ENSInput component
