+++ ui/app/css/itcss/settings/variables.scss
@@ -138,8 +138,14 @@ $Orange-500: #F66A0A;
   font-weight: 400;
 }
 
-%h6 {
+%h5 {
   font-size: 1rem;
+  line-height: 1.25rem;
+  font-weight: 400;
+}
+
+%h6 {
+  font-size: .875rem;
   line-height: 1.25rem;
   font-weight: 400;
 }
